remove setUp from the configuration tests and move it inot baseTest
