+++ java/src/test/java/com/saucelabs/simplesauce/unit/BaseConfigurationTest.java
@@ -1,0 +1,23 @@
+package com.saucelabs.simplesauce.unit;
+
+import com.saucelabs.simplesauce.SauceSession;
+import com.saucelabs.simplesauce.interfaces.EnvironmentManager;
+import com.saucelabs.simplesauce.interfaces.RemoteDriverInterface;
+import org.junit.Before;
+
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+public class BaseConfigurationTest {
+    public SauceSession fakeSauceSession;
+
+    @Before
+    public void setUp()
+    {
+        RemoteDriverInterface fakeRemoteDriver = mock(RemoteDriverInterface.class);
+        EnvironmentManager fakeEnvironmentManager = mock(EnvironmentManager.class);
+        fakeSauceSession = new SauceSession(fakeRemoteDriver, fakeEnvironmentManager);
+        when(fakeEnvironmentManager.getEnvironmentVariable("SAUCE_USERNAME")).thenReturn("test-name");
+        when(fakeEnvironmentManager.getEnvironmentVariable("SAUCE_ACCESS_KEY")).thenReturn("accessKey");
+    }
+}
