+++ java/src/test/java/com/saucelabs/simplesauce/unit/MacOsTests.java
@@ -16,21 +16,7 @@ import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class MacOsTests {
-
-    //TODO duplication in 3 classes, excluding DataCenterTest
-    private SauceSession fakeSauceSession;
-
-    @Before
-    public void setUp()
-    {
-        RemoteDriverInterface fakeRemoteDriver = mock(RemoteDriverInterface.class);
-        EnvironmentManager fakeEnvironmentManager = mock(EnvironmentManager.class);
-        fakeSauceSession = new SauceSession(fakeRemoteDriver, fakeEnvironmentManager);
-        when(fakeEnvironmentManager.getEnvironmentVariable("SAUCE_USERNAME")).thenReturn("test-name");
-        when(fakeEnvironmentManager.getEnvironmentVariable("SAUCE_ACCESS_KEY")).thenReturn("accessKey");
-    }
-
+public class MacOsTests extends BaseConfigurationTest{
     @Test
     public void withMacOsMojave_returnsMacOs1014() throws MalformedURLException {
         fakeSauceSession.withMacOsMojave();
